#18 Changed project structure
