+++ clientapp/src/AppRouter.js
@@ -6,8 +6,8 @@ import { LoginContainer } from './containers/LoginContainer'
 import { RegisterContainer } from './containers/RegisterContainer'
 import PrivateRoute from './PrivateRoute'
 import { ProfileContainer } from './containers/ProfileContainer'
-import { getUser } from './redux/selectors'
-import { refreshTokensRequest } from './redux/actions'
+import { getUser } from './selectors/selectors'
+import { refreshTokensRequest } from './actions'
 import { HomePageContainer } from './containers/HomePageContainer'
 import { SelectedMovieContainer } from './containers/SelectedMovieContainer'
