+++ clientapp/src/components/Profile/ProfileDeleteView.js
