+++ clientapp/src/containers/NavBarContainer.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useHistory } from 'react-router-dom'
 
-import NavBar from '../views/NavBar'
-import { getFullName, getUserAvatar, isLoginSucceeded } from '../redux/selectors'
-import { logoutRequest } from '../redux/actions'
+import NavBar from '../components/NavBar'
+import { getFullName, getUserAvatar, isLoginSucceeded } from '../selectors/selectors'
+import { logoutRequest } from '../actions'
 
 export const NavBarContainer = () => {
   const dispatch = useDispatch()
