+++ clientapp/src/sagas/GenericMovieSaga.js
@@ -1,34 +1,31 @@
 import lodash from 'lodash'
 import { put, select, takeEvery } from 'redux-saga/effects'
 
-import { getJwt } from '../selectors'
+import { getJwt } from '../selectors/selectors'
 import * as callMethods from './callMethods'
-import { enqueueSnackbarError, startLoadingUser, stopLoadingUser } from '../actions'
+import { enqueueSnackbarError } from '../actions'
 
-export function * sagaAllUsers (action) {
-  yield takeEvery(({ type }) => /_REQUEST$/g.test(type), GenericUsersSaga)
+export function * sagaAllMovies (action) {
+  yield takeEvery(({ type }) => /_REQUESTFORMOVIE$/g.test(type), GenericMoviesSaga)
 }
 
-export function * GenericUsersSaga (action) {
+export function * GenericMoviesSaga (action) {
   const { payload, type } = action
   const methodName = lodash.camelCase(type)
   const token = yield select(getJwt)
-  yield put(startLoadingUser())
 
   try {
     const response = yield callMethods[methodName](payload, token)
-    yield put(stopLoadingUser())
-    const successType = action.type.replace('_REQUEST', '_SUCCESS')
+    const successType = action.type.replace('_REQUESTFORMOVIE', '_SUCCESS')
     yield put({ type: successType, payload: response.data })
   } catch (e) {
-    yield put(stopLoadingUser())
     yield put(
       enqueueSnackbarError({
         message: e.response.data.title || e.response.data,
         key: new Date().getTime() + Math.random(),
       })
     )
-    const failedType = action.type.replace('_REQUEST', '_FAIL')
+    const failedType = action.type.replace('_REQUESTFORMOVIE', '_FAIL')
     yield put({ type: failedType, payload: e.response, e })
   }
 }
